# Combine chained `append` filters into a single `append_all`