This PR introduces the `Optimizer` class that can only optimize one thing for now: combine `| append: 'a' | append: 'b'` into a single `| append_all: ['a', 'b']` filter call.

This avoids all the intermediate allocations caused by several `append`.

Also introduce a new filter `append_all: [*items]`. (Maybe I should modify the original `append` instead?)

### Context

We discovered some merchants are using `append` like so:

```
{%- for product in collection.products -%}
  {%- for variant in product.variants -%}
    [... snip ...]
    {% assign prod = productId | append: productSku | append: productTitle | append: productPrice | append: productVendor | append: productCategory | append: productHandle | append: productVariant %}
    [... snip ...]
  {%- endfor -%}
{%- endfor -%}
```

And it's causing big slowdowns, in part because of GC. So this is a first attempt at reducing memory usage.

### Benchmark

With the following code:
```
@context['s'] = "x" * 10240
appends = "| append: 'x' " * 10
Template.parse("{{ s #{appends} }}").render(@context)
```

Results:
```
Warming up --------------------------------------
     with optimizer:     1.021k i/100ms
  without optimizer:   849.000  i/100ms
Calculating -------------------------------------
     with optimizer:     10.199k (± 1.5%) i/s -    102.100k in  10.013380s
  without optimizer:      8.523k (± 2.8%) i/s -     85.749k in  10.069714s

Comparison:
     with optimizer::    10198.7 i/s
  without optimizer::     8522.6 i/s - 1.20x  (± 0.00) slower
```

### Memory

Using `MemoryProfiler` with 1000 runs of code above:

/  | Allocated memory by `standardfilters.rb`
-- | --------------------
with optimizer | 10331000 bytes
without optimizer | 102865000 bytes

About 10x less memory