@macournoyer 2020-10-08
Discussed w/ @tjoyal. We want to avoid adding new filters or modifying existing ones. Maybe it could be possible to optimize using the existing `join` filter instead?