@macournoyer 2020-10-11
Closing in favor of https://github.com/Shopify/liquid-c/pull/82